Warning cleanup and Health graphics changed
Removed many of the warnings in the code and also made the Lives
graphics in the top left corner count down each time you died, just for
use in the future when the player will have multiple lives. Also changed
the old characters code file to a more intuitive named spaceship.